pipeline/vocabulary/requirements.txt
# Bible Vocabulary Pipeline Dependencies

# NLP - Lemmatization
nltk>=3.8

# Utilities
tqdm>=4.66

# Z.AI API (for add_definitions.py --api mode)
zai-sdk==0.0.4.3

# Note: Run this after installing nltk:
# python -c "import nltk; nltk.download('wordnet'); nltk.download('omw-1.4')"
